changelog 17-sept ------ VERSION 3.12
CHANGELOG __init__.py:

added a new feature inside jumptocut called auto-markers (jumptocut.py, functions.py)

basically what auto-markers does is to make IN and OUT markers stable and unremovable from the scene, and also it lets you fine-tune the markers position from the UI...

fixed 'unmeta and trim operator' to align the strips after the operation, and also works fine with multiple strips...

jump to cut new UI

+ keymaps
    -----old ones remains...
    ctrl+shift+right_arrow > skip one second (acording to fps)
    ctrl+shift+left_arrow  < skip one second (acording to fps)
    Q           jump to previous editpoint
    W           jump to next editpoint
    shift+Q     jump to previous marker
    shift+W     jump to next marker
    -----and new ones:
    I           insert IN marker
    O           insert OUT marker

---

CHANGELOG functions.py:

added some scene properties (auto_markers, in_marker and out marker) this values are keeped with each scene in blender when you save the file.

 TODO: recode operator_exta_actions.py cleaning
 or reordering some variables

EXPLANATION: initSceneProperties is ONLY for varaibles that should
be keeped with the blend file. Any other addon preferences
should go to the addon preferences operator in __init__

Many functions from operator_extra_actions.py are using scene properties but should be better to recode it to use this variables as addon preferences.

---

CHANGELOG jumptocut.py:

many changes to use the new auto-markers feature and old behaviour...

a lot of cleanup and labeling in the Jumptocut UI panel...

the "Unmeta and trim" operator now seems to works fine also with multiple selection. It is always a channel-chaos-hazard by using this stuff... but if you use it with enought empty channels avobe the meta then should be no problem...

the most important new piece of code is the bottom function from the file: marker_handler(),
it runs after every scene update and moves the IN marker to the
scene.in_marker value, the same with OUT marker, it also checks everytime that (OUT < IN) never happends, and keeps everything visually refreshed in the VSE...

for the moment it's also a way to test blender using handlers inside the VSE...

---

lots of fixes cleaning bpy. calls around functions.initSceneProperties in many files.